logging: move contents of __init__.py to flask.py
this allows logging.formatting to be imported without implicitly
importing all of these mostly-flask-related logging routines

this is important because we want to be able to use
logging.formatting from the gunicorn config which executes before
eventlet's monkeypatching has been performed, and implicitly
importing flask before monkeypatching adds to the spew of errors
emitted on app startup.
